# openapi.yaml
...
openapi: 3.1.0
info:
  description: Invoices and payments for each business. Amounts are whole US cents.
  title: Invoice and Payment Service
  version: 1.0.0
servers:
- url: /
security:
- apiKey: []
paths:
  /api/v1/customers:
    get:
      operationId: listCustomers
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCustomer"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Customers
    post:
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCustomer"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: This request is too large.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Customers
  /api/v1/customers/{id}:
    get:
      operationId: getCustomer
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Customers
...
  /api/v1/invoices:
    get:
      operationId: listInvoices
      parameters:
      - in: query
        name: state
        required: false
        schema:
          type: string
          enum:
          - open
          - paid
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageInvoiceSummary"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Invoices
    post:
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this customer.
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: This request is too large.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Invoices
  /api/v1/invoices/preview:
    post:
      operationId: previewInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
        required: true
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this customer.
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: This request is too large.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Invoices
  /api/v1/invoices/{id}:
    get:
      operationId: getInvoice
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Invoices
...
components:
  schemas:
    Customer:
      properties:
        created_at:
          type: string
          format: date-time
        email:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
    ErrorDetails:
      properties:
        code:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/InvalidField"
        message:
          type: string
        request_id:
          type: string
    ErrorResponse:
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetails"
...
    InvalidField:
      properties:
        field:
          type: string
        message:
          type: string
    Invoice:
      properties:
        allowed_actions:
          type: array
          items:
            type: string
        amount_display:
          type: string
        created_at:
          type: string
          format: date-time
        currency:
          type: string
        customer_id:
          type: string
          format: uuid
        due_date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        paid_at:
          type: string
          format: date-time
        payment_block_reason:
          type: string
          enum:
          - invoice_already_paid
          - payment_in_progress
        state:
          type: string
          enum:
          - open
          - paid
        total_amount_cents:
          type: integer
          format: int64
    InvoiceInput:
      properties:
        customer_id:
          type: string
          format: uuid
        due_date:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
          maxItems: 100
          minItems: 1
      required:
      - customer_id
      - due_date
      - items
    InvoiceItem:
      properties:
        description:
          type: string
        quantity:
          type: integer
          format: int32
        unit_amount_cents:
          type: integer
          format: int64
        unit_amount_display:
          type: string
    InvoiceSummary:
      properties:
        amount_display:
          type: string
        created_at:
          type: string
          format: date-time
        currency:
          type: string
        customer_id:
          type: string
          format: uuid
        due_date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        paid_at:
          type: string
          format: date-time
        state:
          type: string
          enum:
          - open
          - paid
        total_amount_cents:
          type: integer
          format: int64
    LineItem:
      properties:
        description:
          type: string
          maxLength: 500
          minLength: 1
        quantity:
          type: integer
          format: int32
          maximum: 10000
          minimum: 1
        unit_amount_cents:
          type: integer
          format: int64
          maximum: 1000000000000
          minimum: 0
      required:
      - description
      - quantity
      - unit_amount_cents
    NewCustomer:
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          minLength: 1
        name:
          type: string
          maxLength: 200
          minLength: 1
      required:
      - email
      - name
...
    PageCustomer:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        next_cursor:
          type: string
...
    PageInvoiceSummary:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceSummary"
        next_cursor:
          type: string
...
  securitySchemes:
    apiKey:
      description: Use a valid API key.
      scheme: bearer
      type: http